feat: add login controller
