+++ src/routes/accounts/login.ts
@@ -1,9 +1,6 @@
-import { Request, Response } from 'express';
 import { body } from 'express-validator';
-import jwt from 'jsonwebtoken';
 
-import { User } from '../../models';
-import { PasswordManager } from '../../util/hash/index';
+import { loginHandler } from '../../controller/account/login.controller';
 
 const loginRouter = require('express').Router();
 
@@ -14,41 +11,7 @@ loginRouter.post(
     body('email').isLength({ max: 30 }).isEmail().escape(),
     body('password').isLength({ min: 2, max: 20 }).escape(),
   ],
-  async (req: Request, res: Response) => {
-    const reqbody = req.body;
-    // Find User in Database
-    const user = await User.findOne({ username: reqbody.username });
-    const passwordCorrect =
-      user === null
-        ? false
-        : await PasswordManager.compare(user.password, reqbody.password);
-    // Return error is user does not exist or password is incorrect
-    if (!(user && passwordCorrect)) {
-      res.status(401).json({
-        error: 'invalid username or password',
-      });
-    }
-    // username & password are correct
-    else {
-      // create a jwt
-      const userJwt = jwt.sign(
-        {
-          id: user.id,
-          email: user.email,
-          username: user.username,
-        },
-        process.env.JWT_KEY!,
-        { expiresIn: 60 * 60 }
-      );
-      res
-        .status(200)
-        .cookie('access_token', userJwt, {
-          httpOnly: true,
-          secure: process.env.NODE_ENV === 'production',
-        })
-        .json(user);
-    }
-  }
+  loginHandler
 );
 
 export { loginRouter };
