+++ src/routes/login.ts
@@ -1,10 +1,25 @@
+/* eslint-disable no-throw-literal */
 import { Request, Response } from 'express';
 import { body } from 'express-validator';
 import jwt from 'jsonwebtoken';
+
 import { User } from '../models';
 import { PasswordManager } from '../util/hash/index';
 
 const loginRouter = require('express').Router();
+
+interface MyError {
+  status: number;
+  message: string;
+}
+
+const customError = (
+  status: number,
+  message: string = 'Invalid credentials'
+): MyError => ({
+  status,
+  message,
+});
 
 loginRouter.post(
   '/api/v1/users/login',
@@ -14,21 +29,19 @@ loginRouter.post(
     body('password').isLength({ min: 2, max: 20 }).escape(),
   ],
   async (req: Request, res: Response) => {
-    const reqbody = req.body;
-    // Find User in Database
-    const user = await User.findOne({ username: reqbody.username });
-    const passwordCorrect =
-      user === null
-        ? false
-        : await PasswordManager.compare(user.password, reqbody.password);
-    // Return error is user does not exist or password is incorrect
-    if (!(user && passwordCorrect)) {
-      res.status(401).json({
-        error: 'invalid username or password',
-      });
-    }
-    // username & password are correct
-    else {
+    try {
+      const { username, password } = req.body;
+      // Find User in Database
+      const user = await User.findOne({ username });
+      if (!user) throw customError(401);
+
+      const passwordCorrect = await PasswordManager.compare(
+        user.password,
+        password
+      );
+      // Return error is user does not exist or password is incorrect
+      if (!passwordCorrect) throw customError(401);
+      // username & password are correct
       // create a jwt
       const userJwt = jwt.sign(
         {
@@ -46,6 +59,10 @@ loginRouter.post(
           secure: process.env.NODE_ENV === 'production',
         })
         .json(user);
+    } catch (err: any) {
+      res.status(err.status).json({
+        message: err.message,
+      });
     }
   }
 );
